Removed dev dependency on ParameterHandler

+++ bin/test.sh
@@ -1,7 +1,20 @@
 #!/usr/bin/env bash
 
+PROJECT_ROOT="$( cd "$( dirname "${BASH_SOURCE[0]}" )/.." && pwd )"
+
 echo ''
 echo '// Building test environment'
+
+if [ ! -f $PROJECT_ROOT/tests/App/config/parameters.yaml ]; then
+    unset PROJECT_ROOT
+
+    echo ''
+    echo ' [FAIL] You must configure database in tests/App/config/parameters.yaml with be able to run the tests'
+    echo ''
+
+    exit 1
+fi
+unset PROJECT_ROOT
 
 rm -rf tests/App/var/cache/* tests/App/var/log/*
 composer --quiet --no-interaction update --optimize-autoloader > /dev/null
